Move KeyPairHolder* to core-kernel
