+++ packages/core-forger/__tests__/delegate-tracker.test.ts
@@ -2,8 +2,8 @@ import "jest-extended";
 
 import { Delegate } from "@packages/core-forger/src/delegate";
 import { DelegateTracker } from "@packages/core-forger/src/delegate-tracker";
-import { Bip39KeyPairHolder } from "@packages/core-forger/src/key-pair-holders/bip39-key-pair-holder";
 import { Utils } from "@packages/core-kernel";
+import { Bip39KeyPairHolder } from "@packages/core-kernel/src/utils/key-pair-holders/bip39-key-pair-holder";
 import { Wallet } from "@packages/core-state/src/wallets";
 import { Crypto, Managers } from "@packages/crypto";
