+++ packages/core-kernel/src/utils/index.ts
@@ -12,10 +12,11 @@ export * as Search from "./search";
 import { calculate } from "./supply-calculator";
 
 export * from "@arkecosystem/utils";
-export * from "./expiration-calculator";
 export * from "./assert";
+export * from "./expiration-calculator";
 export * from "./ipc-handler";
 export * from "./ipc-subprocess";
+export * from "./key-pair-holders";
 export * from "./lock";
 
 export const delegateCalculator = { calculateApproval, calculateForgedTotal };
